moved room component to its own folder
