+++ project/src/components/app/app.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import MainPage from '../main-page';
 import SignIn from '../sign-in';
 import Favorites from '../favorites';
-import Room from '../room';
+import Room from '../room/room';
 import PrivateRoute from './private-route';
